Create abstraction.py
It goes over a jail file and figures out if two handlers are essentially the same.
If they are the same, then it prints them out.

If they aren't the same, but they're at least similar in length (so they're at least
approximately the same), then it outputs a diff of the two python objects representing
the handlers to a file in a folder called diffs/. The diffs/ folder is in the .gitignore
so that you don't end up adding a bunch of .html files to the repo.

In addition, the output is pretty-printed using the json python module, so the diffs
look really great.