Create program to find missing projects, assrt jail2 updates
I made findmissing.py to find out if I was missing any directories, and to check on my progress with fixing jail2.py.

I made several updates to jail2.py such as making it show you an actual difference between two times, because it was actually showing the current time instead. This was because logStartTime was set to 0 initially, and the log file init was never called.

It also actually sets currentScreenName properly now, I believe.