Add docs to jail2, find missing files
I added more documentation of changes to jail2.py, these changes were mostly variable name changes.

I also created a function that finds files that are present in the aias but are not present in the jail output, so that I know which projects are the 'problem children' so to speak.