Attempt to fix types
There is a lot of *crap*, shall we say, going on with types at the moment. All of the types seem to be incorrect for one reason or another, and I'm just trying to go piece by piece and fix it. I might want to shift away from this and try to fix something that will at least get rid of more errors, such as the problem with form name and screen names not matching.