Apparently the copy of abstraction.py is open on another branch
Usually the editors that I use change which branch I'm on automatically, but it doesn't seem that way for emacs, for whatever reason. I've been trying to wrap my head around why abstraction.py can't find more than one screen for 09265's project bake, but it turns out it's on the wrong branch?