Create equivalence classes for abstraction.py
I made actual equivalence classes as a class in python and then created a CodeSet class that contains all of the equivalence classes found in the code for a given screen. These classes manage some of the complexity of testing for whether or not a piece of code already belongs to an equivalence class, which is extremely handy. I'm pretty sure that I got the logic right too, which is nice.

I have to admit I really miss object oriented programming though lol.