Adapt abstraction to work with zips
I was able to get abstraction.py to work with the zip files, and it actually was relatively fast to analyze them -- about 12 and a half minutes for 1mil, so it'll likely be about 18-19 minutes for 1.5mil.

I just need to find a way to either save these things, or maybe some other way of saving things. I also need to modify the different equivalence class things so that tehy will work better with screen names, because at the moment, I'm actually not sure which screen each handler comes from. I also only had it analyze declaration type top blocks, which may have made it faster.