Cut 46k findmissing.py exec time to 10min
I did so by removing items in the jailFiles array that we'd already found the correspondents to from the aiaFiles array. So at first, it started out taking 3 seconds to do check 5000 files, but this was way better than the 5 seconds it started to take the original compareFiles function to do 5000. It also sped up in a probably linear fashion throughout the entire run, which was good.

I'm also working on finding out what the disabled tag is.