Make # of jail files consistent
Apparently there are 30,852 things in the resulting equivs array, but there were only 30,851 jail files found in the first place. And yet, my findmissing program claims that there are 134 files missing from the jail files, when there are 30,984 aia files, and 30,851 jail files, so there should only be 133 missing, and yet there are 134 missing somehow. I'll eventually get to the bottom of this.