Add function for determining whether needs generics
I added a function to the equivalence classes to determine whether the equivalence class would need generics or not to fully abstract over. This function is EquivalenceClass.needsGenerics().

I also added code to abtsraction.py for outputting information about this to standard output as well as to a CSV.

myutils.py was changed in order to add a function that would count how many of the blocks in a handler are generics, countGenerics([some block]).