Slowly moving away from coding like a terrible person
I have a bunch of 'libraries' that are really just swaths of code that I made that do certain utility functions, and I was importing all of each 'library' into abstraction.py, which is just really terrible for namespaces and such. Apparently it works, and puts out the correct output still. In this particular instance, I was preventing it from importing all of myutils, and just imported myutils as mus.