Improve refactoring, found color discrepancy
Still working on refactoring a bit, although I've also been actually using loopanalysis.py to actually find things. I also found a discrepancy in the colors that are reported in jail as the type versus what the actual value of the color is, and wrote code to change that.

I think there's some better separation overall of the code that is specific to abstraction.py and code that is just helpers, and I've been fairly good about making my helpers in myutils.py instead of in loopanalysis itself, as well as file io stuff being put in findmissing.py, which should probably be renamed.